Fixed up Sector and moved SectorTexture
